Change pinch detection
Just two-finger-pinches are now detected but not triggered directly.
Instead they are stored and only after a certain amount of time they are
triggered. During that time, other detected pinches may be detected as
part of previous pinches and therefore merged.

Before multi-finger-pinches had to start at the exact same time but
sometimes one of the sub-pinches starts early.